patch: avoid dropping calls while negociating direct_media with update
As soon as 200 OK of initial transaction has been sent, the direct_media
process kicks in. If direct_media_method is 'update', while receiving the
ACK to the initial transaction, the source was checking that the SDP
negociation was complete.

As the UPDATE has already been sent, this is not the case, so the call was
being dropped.

This problem doesn't happen with direct_media_method 'invite' becuase the
initial transaction must be completed before initiating direct_media
process.

This patch updates the conditions to also consider as valid a pending remote
negociation response, so asterisk will have time to handle the UPDATE response.